Templates and views for the django site
urls has been changed to get the site working on the board.html
template as the index site.
views has been change and a board function has been added to return the
board.html page.
html tamplate for basic site with the data model

** Every line in the site will look like this:
   posted by <username> at <time>
   <user data>
** At the end of the site "(End of messages)" will be added.

*** in the views.py there are a few lines that are refering to POST
request it is an improvment on this commit and will be explained in the
next PR